pyproject.toml
[project]
name = "Job_Platform"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "vincent", email = "vincent992538@gmail.com"},
]
dependencies = ["django>=5.2.1", "django-ninja>=1.4.1", "django-ninja-jwt>=5.3.7", "pytest-asyncio>=0.26.0"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "httpx>=0.28.1",
    "flake8>=7.2.0",
]
